man: Migrate to doc/
As sphinx supports generating manpages, use it to simplify the process
and drop the dependency on rst2man.

All manpages are now tracked in doc/man.

The Makefile in man/ remains for forward compatibility.

Keep tracking manpages themselves in troff format so sphinx is not
required for packaging, slightly update Documentation to reflect the
changes requirements for manpage generation.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>